Fix Esc key detection and close issue #6
Implements solution for differentiating between special characters
preceded by escape key and the actual escape key itself. Initially
a blocking read is performed to read a character and then if the input
buffer still conains characters they are read in non-blocking mode.
This means that if a user presses only escape key then the buffer will
be empty and there will not be any more bytes to read.

This solution is safe for reading multibyte characters as they are going
to be handled by regular blocking read and any bytes following the escape
key will not contain multibyte.