Overload nonblock semantics with a timeout value
Useful if you have an application that would like to sleep at the end
of every loop. This will let it wait for & collect the keypress until the
sleep would have expired.

For example, this loop will sleep until approximately the start of the next
second on the system clock.

```
loop do
  draw
  now = Time.now.to_f
  codes = reader.read_keypress(raw: true, nonblock: (now.ceil(0) - now))
  command.interpret(codes) if codes
end
```